#Codefoces_1566A[A. Median Maximization]
A. Median Maximization
time limit per test: 1 second
memory limit per test: 256 megabytes
input: standard input
output: standard output

You are given two positive integers and . Find the maximum possible median of an array of non-negative integers (not necessarily distinct), such that the sum of its elements is equal to .
A median of an array of integers of length m is the number standing on the[m/2] -th (roundingup) position in the non-decreasing ordering of its elements. Positions are numbered starting from 1. For example, a median of the array [20, 40, 20, 50, 50, 30] is the [m/2]-th element of [20, 20, 30, 40, 50, 50], so it is 30 . There exist other definitions of the median, but in this problem we use the described definition.

Input
The input consists of multiple test cases. The first line contains a single integer t.
( 1<= t <= 10^4 )— the number of test cases. Description of the test cases follows.
Each test case contains a single line with two integers  'n' and 's'( ) — the length
of the array and the required sum of the elements.
Output
For each test case print a single integer — the maximum possible median.